+++ src/main/java/Controladores/Creditos/controladorCL.java
@@ -6,6 +6,7 @@ package Controladores.Creditos;
 
 import Modelos.Creditos.Credito;
 import Modelos.Creditos.CreditoLibre;
+import java.util.ArrayList;
 
 /**
  *
@@ -14,18 +15,51 @@ import Modelos.Creditos.CreditoLibre;
 public class ControladorCL extends ControladorCredito {
 
   
-    public ControladorCL(Credito credito ){
-        super(credito);
+    private ArrayList<CreditoLibre> creditosLibres;
+    
+    public ControladorCL(){
+        creditosLibres = new ArrayList<>();
     }
+    
+    /**
+     * Metodo para saber si un credito de libre inversion ya está almacenado
+     * @param numeroContrato
+     * @return credito de libre inversion si ya esta almacenado, de lo contrario null
+     */
+    public CreditoLibre buscarCreditoL( int numeroContrato ){
+        
+        for (CreditoLibre creditoL : creditosLibres) {
+            if( creditoL.getNumeroDeContrato() == numeroContrato ) return creditoL;
+        }
+        
+        return null;
+    }
+    
+    /**
+     * Metodo para añadir un credito libre
+     * @param creditoL
+     * @return true si se añadió, de lo contrario false
+     */
+    public boolean añadirCreditoL(CreditoLibre creditoL){
+        CreditoLibre aux = buscarCreditoL(creditoL.getNumeroDeContrato());
+        
+        if( aux == null ){
+            creditosLibres.add(creditoL);
+            return true;
+        }
+        
+        return false;
+    }
+    
     
     /**
      * Metodo para calcular el valor de la cuota del credito de libre expresion
      * @return 
      */
     @Override
-    public void calcularValorCuota(){
-        CreditoLibre creditoL = this.cB.buscarCreditoL(this.getNumeroContrato());
-        double soli = this.getValorSolicitado();
+    public void calcularValorCuota(Credito credito){
+        CreditoLibre creditoL = buscarCreditoL(credito.getNumeroDeContrato());
+        double soli = creditoL.getMontoTotal();
         int valorData = creditoL.getValorDataCredito();
         double valorTotal = creditoL.getValorCuota();
         
@@ -36,6 +70,20 @@ public class ControladorCL extends ControladorCredito {
         
         valorTotal+=soli;
         creditoL.setValorCuota(valorTotal);
+    }
+
+    /**
+     * @return the creditosLibres
+     */
+    public ArrayList<CreditoLibre> getCreditosLibres() {
+        return creditosLibres;
+    }
+
+    /**
+     * @param creditosLibres the creditosLibres to set
+     */
+    public void setCreditosLibres(ArrayList<CreditoLibre> creditosLibres) {
+        this.creditosLibres = creditosLibres;
     }
     
 }
