+++ src/main/java/Controladores/Creditos/controladorCredito.java
@@ -4,7 +4,6 @@
  */
 package Controladores.Creditos;
 
-import Controladores.ControladorBanco;
 import Modelos.Creditos.Credito;
 import Modelos.Cuota.Cuota;
 
@@ -13,25 +12,14 @@ import Modelos.Cuota.Cuota;
  * @author USER
  */
 public abstract class ControladorCredito {
-    protected Cuota[] cuotas;
-    protected Credito credito;
-    protected double valorSolicitado;
-    protected int numeroContrato;
-    protected ControladorBanco cB;
-    
-    public ControladorCredito( Credito credito ){
-        this.credito = credito;
-        cuotas = credito.getCuotas();
-        this.valorSolicitado = credito.getMontoTotal();
-        this.numeroContrato = credito.getNumeroDeContrato();
-        cB = new ControladorBanco();
+    public ControladorCredito( ){
     }
     
     
     /**
      * Metodo para calcular el valor total de la cuota
      */
-    public abstract void calcularValorCuota();
+    public abstract void calcularValorCuota(Credito credito);
     
     
     /**
@@ -39,11 +27,13 @@ public abstract class ControladorCredito {
      * @param cuota
      * @return true si se abonó, de lo contrario false
      */
-    public boolean abonarCuota(Cuota cuota){
+    public boolean abonarCuota(Credito credito, Cuota cuota){
         
         //LOOK OUT THIS PART
-        if( cuota.getMontoAbonado() < this.credito.getValorCuota() ) return false;
+        if( cuota.getMontoAbonado() < credito.getValorCuota() ) return false;
         int contador = 0;
+        Cuota[] cuotas = credito.getCuotas();
+        
         for (int i = 0; i < cuotas.length; i++) {
             if( cuotas[i] != null ) contador++;
             
@@ -61,9 +51,10 @@ public abstract class ControladorCredito {
      * Metodo para saber la cantidad de cuotas restantes
      * @return la cantidad de cuotas restantes
      */
-    public void getCuotasRestantes(){
+    public void getCuotasRestantes(Credito credito){
         int contadorAbonadas = 0;
         int contadorRestantes = 0;
+        Cuota[] cuotas = credito.getCuotas();
         for (int i = 0; i < cuotas.length; i++) {
             if( cuotas[i] == null ){
                 contadorRestantes++;
@@ -71,49 +62,8 @@ public abstract class ControladorCredito {
                 contadorAbonadas++;
             }
         }
-        this.credito.setCuotasRestantes(contadorRestantes);
-        this.credito.setCuotasAbonadas(contadorRestantes);
+        credito.setCuotasRestantes(contadorRestantes);
+        credito.setCuotasAbonadas(contadorRestantes);
     }
 
-    /**
-     * @return the cuotas
-     */
-    public Cuota[] getCuotas() {
-        return cuotas;
-    }
-
-    /**
-     * @param cuotas the cuotas to set
-     */
-    public void setCuotas(Cuota[] cuotas) {
-        this.cuotas = cuotas;
-    }
-
-    /**
-     * @return the valorSolicitado
-     */
-    public double getValorSolicitado() {
-        return valorSolicitado;
-    }
-
-    /**
-     * @param valorSolicitado the valorSolicitado to set
-     */
-    public void setValorSolicitado(double valorSolicitado) {
-        this.valorSolicitado = valorSolicitado;
-    }
-
-    /**
-     * @return the numeroContrato
-     */
-    public int getNumeroContrato() {
-        return numeroContrato;
-    }
-
-    /**
-     * @param numeroContrato the numeroContrato to set
-     */
-    public void setNumeroContrato(int numeroContrato) {
-        this.numeroContrato = numeroContrato;
-    }
 }
