+++ src/main/java/Controladores/controladorBanco.java
@@ -4,8 +4,6 @@
  */
 package Controladores;
 
-import Modelos.Creditos.CreditoHipotecario;
-import Modelos.Creditos.CreditoLibre;
 import Modelos.Datos.Cliente;
 import java.util.ArrayList;
 
@@ -14,14 +12,10 @@ import java.util.ArrayList;
  * @author USER
  */
 public class ControladorBanco {
-    private ArrayList<CreditoHipotecario> creditosHipotecarios;
-    private ArrayList<CreditoLibre> creditosLibres;
     private ArrayList<Cliente> clientes;
     //private ArrayList<Domicilio> domicilios;
     
     public ControladorBanco(){
-        creditosHipotecarios = new ArrayList<>();
-        creditosLibres = new ArrayList<>();
         clientes = new ArrayList<>();
         //domicilios = new ArrayList<>();
     }
@@ -119,97 +113,17 @@ public class ControladorBanco {
     /**************** GESTIONAR CREDITOS *******************/
     
     
-    /*
-    *****  CREDITO HIPOTECARIO
-    *creditoH se usa como abreviacion de credito Hipotecario
-    */
     
     
-    /**
-     * Metodo para saber si un credito hipotecario ya está almacenado
-     * @param numeroContrato
-     * @return credito hipotecario si ya esta almacenado, de lo contrario null
-     */
-    public CreditoHipotecario buscarCreditoH( int numeroContrato ){
-        
-        for (CreditoHipotecario creditoH : creditosHipotecarios) {
-            if( creditoH.getNumeroDeContrato() == numeroContrato ) return creditoH;
-        }
-        
-        return null;
-    }
-    
-    /**
-     * Metodo para añadir un credito hipotecario
-     * @param creditoH
-     * @return true si se añadió, de lo contrario false
-     */
-    public boolean añadirCreditoH(CreditoHipotecario creditoH){
-        CreditoHipotecario aux = buscarCreditoH(creditoH.getNumeroDeContrato());
-        
-        if( aux == null ){
-            creditosHipotecarios.add(creditoH);
-            return true;
-        }
-        
-        return false;
-    }
     
     
-    /*
-    *****  CREDITO Libre
-    *creditoL se usa como abreviacion de credito de libre inversion
-    */
-    
-    
-    /**
-     * Metodo para saber si un credito de libre inversion ya está almacenado
-     * @param numeroContrato
-     * @return credito de libre inversion si ya esta almacenado, de lo contrario null
-     */
-    public CreditoLibre buscarCreditoL( int numeroContrato ){
-        
-        for (CreditoLibre creditoL : creditosLibres) {
-            if( creditoL.getNumeroDeContrato() == numeroContrato ) return creditoL;
-        }
-        
-        return null;
-    }
-    
-    /**
-     * Metodo para añadir un credito libre
-     * @param creditoL
-     * @return true si se añadió, de lo contrario false
-     */
-    public boolean añadirCreditoL(CreditoLibre creditoL){
-        CreditoHipotecario aux = buscarCreditoH(creditoL.getNumeroDeContrato());
-        
-        if( aux == null ){
-            creditosLibres.add(creditoL);
-            return true;
-        }
-        
-        return false;
-    }
     
     
     
     
     /**************** GETTERS *******************/
     
-    /**
-     * @return the creditosHipotecarios
-     */
-    public ArrayList<CreditoHipotecario> getCreditosHipotecarios() {
-        return creditosHipotecarios;
-    }
-
-    /**
-     * @return the creditosLibres
-     */
-    public ArrayList<CreditoLibre> getCreditosLibres() {
-        return creditosLibres;
-    }
+    
 
     /**
      * @return the clientes
